Add files via upload
a program that simulates the game of craps, which is played with two dice. On the first roll, the player wins if the sum of the dice is 7 or II. The player loses if the sum is 2, 3, or 12. Any other roll is called the "point" and the game continues. On each subsequent roll, the player wins if he or she rolls the point again. The player loses by rolling 7. Any other roll is ignored and the game continues. At the end of each game, the program will ask die user whether or not to play again. When the user enters a response other than y or Y. the program will display the number of wins and losses and then terminate.